+++ harpy/__main__.py
@@ -3,16 +3,13 @@
 import rich_click as click
 from harpy.commands import align
 from harpy.commands import assembly, metassembly
-from harpy.commands import convert
 from harpy.commands import diagnose, resume, view
 from harpy.commands import deconvolve
 from harpy.commands import demultiplex
-from harpy.commands import downsample
 from harpy.commands import environments
 from harpy.commands import impute
 from harpy.commands import qc
 from harpy.commands import phase
-from harpy.commands import simulate
 from harpy.commands import snp, sv
 from harpy.commands import template
 from harpy.commands import validate
@@ -38,12 +35,12 @@ config = click.RichHelpConfiguration(
 @click.command_panel(
     "Data Processing",
     panel_styles={"border_style": "blue"},
-    commands = sorted(["align","demultiplex","qc","snp","sv","impute","phase", "simulate", "assembly", "metassembly"])
+    commands = sorted(["align", "deconvolve", "demultiplex","qc","snp","sv","impute","phase", "assembly", "metassembly"])
 )
 @click.command_panel(
     "Other Commands",
     panel_styles={"border_style": "magenta"},
-    commands = sorted(["convert","deconvolve", "downsample", "template"])
+    commands = sorted(["template"])
 )
 @click.command_panel(
     "Troubleshoot",
@@ -52,7 +49,7 @@ config = click.RichHelpConfiguration(
 )
 def cli():
     """
-    An automated workflow for linked-read data
+    Automated workflows for linked-read data
     to go from raw data to genotypes (or phased haplotypes).
     Batteries included.
     
@@ -64,11 +61,9 @@ def cli():
 # main program
 cli.add_command(align.align)
 cli.add_command(assembly.assembly)
-cli.add_command(convert.convert)
 cli.add_command(deconvolve.deconvolve)
 cli.add_command(demultiplex.demultiplex)
 cli.add_command(diagnose.diagnose)
-cli.add_command(downsample.downsample)
 cli.add_command(environments.containerize)
 cli.add_command(environments.deps)
 cli.add_command(impute.impute)
@@ -76,7 +71,6 @@ cli.add_command(metassembly.metassembly)
 cli.add_command(phase.phase)
 cli.add_command(qc.qc)
 cli.add_command(resume.resume)
-cli.add_command(simulate.simulate)
 cli.add_command(snp.snp)
 cli.add_command(sv.sv)
 cli.add_command(validate.validate)